# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Golf Charity Subscription Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /scores:
    get:
      operationId: listMyScores
      tags: [scores]
      summary: Get current user's scores (newest first)
      responses:
        "200":
          description: List of scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Score"

    post:
      operationId: addScore
      tags: [scores]
      summary: Add a golf score (removes oldest if 6th)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddScoreBody"
      responses:
        "201":
          description: Score added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Score"
  /scores/{id}:
    patch:
      operationId: updateScore
      tags: [scores]
      summary: Update a score
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateScoreBody"
      responses:
        "200":
          description: Score updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Score"

    delete:
      operationId: deleteScore
      tags: [scores]
      summary: Delete a score
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Score deleted
  /scores/user/{userId}:
    get:
      operationId: getUserScores
      tags: [scores]
      summary: Get scores for a user (admin)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User scores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Score"

  # Charities
...
  /draws:
    get:
      operationId: listDraws
      tags: [draws]
      summary: List all draws
      responses:
        "200":
          description: List of draws
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Draw"

    post:
      operationId: createDraw
      tags: [draws]
      summary: Create a draw (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDrawBody"
      responses:
        "201":
          description: Draw created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Draw"
  /draws/latest:
    get:
      operationId: getLatestDraw
      tags: [draws]
      summary: Get the most recent published draw
      responses:
        "200":
          description: Latest draw with results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DrawWithResults"
        "404":
          description: No draws yet
  /draws/{id}:
    get:
      operationId: getDraw
      tags: [draws]
      summary: Get draw by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Draw
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DrawWithResults"
  /draws/{id}/simulate:
    post:
      operationId: simulateDraw
      tags: [draws]
      summary: Simulate a draw (preview, admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Simulation results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DrawSimulationResult"
  /draws/{id}/publish:
    post:
      operationId: publishDraw
      tags: [draws]
      summary: Publish/execute the final draw (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Draw published with results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DrawWithResults"

  # Winners
...
components:
  schemas:
...
    Score:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        value:
          type: integer
        scoreDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - value
        - scoreDate
        - createdAt
    AddScoreBody:
      type: object
      properties:
        value:
          type: integer
        scoreDate:
          type: string
          format: date-time
      required:
        - value
        - scoreDate
    UpdateScoreBody:
      type: object
      properties:
        value:
          type: integer
        scoreDate:
          type: string
          format: date-time
...
    Draw:
      type: object
      properties:
        id:
          type: integer
        month:
          type: integer
        year:
          type: integer
        status:
          type: string
        prizePool:
          type: number
        jackpotRollover:
          type: number
        drawnNumbers:
          type: ["string", "null"]
        algorithm:
          type: string
        publishedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - month
        - year
        - status
        - prizePool
        - jackpotRollover
        - algorithm
        - createdAt
    CreateDrawBody:
      type: object
      properties:
        month:
          type: integer
        year:
          type: integer
        algorithm:
          type: string
        jackpotRollover:
          type: number
      required:
        - month
        - year
    DrawResult:
      type: object
      properties:
        id:
          type: integer
        drawId:
          type: integer
        userId:
          type: integer
        matchTier:
          type: integer
        prizeAmount:
          type: number
        userName:
          type: string
        userEmail:
          type: string
      required:
        - id
        - drawId
        - userId
        - matchTier
        - prizeAmount
        - userName
        - userEmail
    DrawWithResults:
      type: object
      properties:
        id:
          type: integer
        month:
          type: integer
        year:
          type: integer
        status:
          type: string
        prizePool:
          type: number
        jackpotRollover:
          type: number
        drawnNumbers:
          type: ["string", "null"]
        algorithm:
          type: string
        publishedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        results:
          type: array
          items:
            $ref: "#/components/schemas/DrawResult"
      required:
        - id
        - month
        - year
        - status
        - prizePool
        - jackpotRollover
        - algorithm
        - createdAt
        - results
    DrawSimulationResult:
      type: object
      properties:
        drawnNumbers:
          type: array
          items:
            type: integer
        tier5Winners:
          type: array
          items:
            $ref: "#/components/schemas/DrawResult"
        tier4Winners:
          type: array
          items:
            $ref: "#/components/schemas/DrawResult"
        tier3Winners:
          type: array
          items:
            $ref: "#/components/schemas/DrawResult"
        prizePool:
          type: number
      required:
        - drawnNumbers
        - tier5Winners
        - tier4Winners
        - tier3Winners
        - prizePool
...