# lib/api-zod/src/generated/types/draw.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Golf Charity Subscription Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Draw {
  id: number;
  month: number;
  year: number;
  status: string;
  prizePool: number;
  jackpotRollover: number;
  /** @nullable */
  drawnNumbers?: string | null;
  algorithm: string;
  /** @nullable */
  publishedAt?: Date | null;
  createdAt: Date;
}