# lib/api-zod/src/generated/types/drawSimulationResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Golf Charity Subscription Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { DrawResult } from "./drawResult";

export interface DrawSimulationResult {
  drawnNumbers: number[];
  tier5Winners: DrawResult[];
  tier4Winners: DrawResult[];
  tier3Winners: DrawResult[];
  prizePool: number;
}