# lib/api-zod/src/generated/types/updateScoreBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Golf Charity Subscription Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateScoreBody {
  value?: number;
  scoreDate?: Date;
}